Rename Classification to Classification.md
I converted the problem to a classification problem to check if Liberty Mutual decides to revisit those properties which are non-hazardous so as to be double sure of the property. This will help Liberty Mutual Insurance to identify high risk homes that require additional examination. We check for various models such as logistic regression, xgboost and random forest and calculate the misclassification rate. All methods are compared using ROC curves and AUC plots.
XGBoost performs best